# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DataPulse Backend API",
    "description": "DataPulse Backend API powering crowdsourced dataset uploads, automated quality evaluation, vector search, and credits redemption.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/datasets/upload": {
      "post": {
        "tags": [
          "Datasets"
        ],
        "summary": "Upload Dataset",
        "operationId": "upload_dataset_datasets_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_datasets_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "List Datasets",
        "operationId": "list_datasets_datasets_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_id}/evaluation": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get Dataset Evaluation",
        "operationId": "get_dataset_evaluation_datasets__dataset_id__evaluation_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "tags": [
          "Datasets"
        ],
        "summary": "Search Datasets",
        "operationId": "search_datasets_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/search": {
      "post": {
        "tags": [
          "Datasets"
        ],
        "summary": "Search Datasets",
        "operationId": "search_datasets_datasets_search_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Search Datasets Get",
        "operationId": "search_datasets_get_datasets_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_dataset_datasets_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "general"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_datasets_upload_post"
      },
      "DatasetDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "rows_count": {
            "type": "integer",
            "title": "Rows Count"
          },
          "columns_count": {
            "type": "integer",
            "title": "Columns Count"
          },
          "schema_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schema Json"
          },
          "content_summary": {
            "type": "string",
            "title": "Content Summary"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "filename",
          "domain",
          "rows_count",
          "columns_count",
          "content_summary",
          "created_at"
        ],
        "title": "DatasetDetailResponse"
      },
      "DatasetListResponse": {
        "properties": {
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetDetailResponse"
            },
            "type": "array",
            "title": "Datasets"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "datasets",
          "total"
        ],
        "title": "DatasetListResponse"
      },
      "EvaluationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Evaluation status: pending, processing, completed, failed"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "description": "Progress percentage from 0 to 100"
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score",
            "description": "Overall quality score 0-100"
          },
          "schema_fidelity_score": {
            "type": "number",
            "title": "Schema Fidelity Score",
            "description": "Schema fidelity score 0-100"
          },
          "completeness_score": {
            "type": "number",
            "title": "Completeness Score",
            "description": "Completeness score 0-100"
          },
          "domain_relevance_score": {
            "type": "number",
            "title": "Domain Relevance Score",
            "description": "Domain relevance score 0-100"
          },
          "overall_score": {
            "type": "number",
            "title": "Overall Score",
            "description": "Composite score 0-100"
          },
          "credits_awarded": {
            "type": "integer",
            "title": "Credits Awarded",
            "description": "Credits awarded for this dataset"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "status",
          "progress",
          "quality_score",
          "schema_fidelity_score",
          "completeness_score",
          "domain_relevance_score",
          "overall_score",
          "credits_awarded",
          "updated_at",
          "created_at"
        ],
        "title": "EvaluationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchItem": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "rows_count": {
            "type": "integer",
            "title": "Rows Count"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "name",
          "domain",
          "score",
          "summary",
          "rows_count"
        ],
        "title": "SearchItem"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Search query string"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain",
            "description": "Optional domain filter"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Max results",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchItem"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "Search algorithm used (embeddings or tfidf_hybrid)"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "total",
          "method"
        ],
        "title": "SearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}